update cache importing to separate file
