+++ deps.ts
@@ -1,7 +1,9 @@
 export * from "https://deno.land/std@0.67.0/fmt/colors.ts";
 export * from "https://deno.land/std@0.67.0/async/delay.ts";
-export * from "https://x.nest.land/Discordeno@9.0.10/mod.ts";
-export * from "https://x.nest.land/Sabr@1.1.0/mod.ts";
+export * from "https://deno.land/x/discordeno@9.0.15/mod.ts";
+export * from "https://deno.land/x/sabr@1.1.0/mod.ts";
+export * from "./cache.ts";
 
-import Client from "https://x.nest.land/Discordeno@9.0.10/src/module/client.ts";
+import Client from "https://deno.land/x/discordeno@9.0.15/src/module/client.ts";
+
 export default Client;
