+++ src/arguments/guild.ts
@@ -1,5 +1,4 @@
-import { botCache } from "../../mod.ts";
-import { cache } from "../../deps.ts";
+import { botCache, cache } from "../../deps.ts";
 
 botCache.arguments.set("guild", {
   name: "guild",
