+++ src/arguments/role.ts
@@ -1,4 +1,4 @@
-import { botCache } from "../../mod.ts";
+import { botCache, cache } from "../../deps.ts";
 import { sendResponse } from "../utils/helpers.ts";
 
 botCache.arguments.set("role", {
@@ -7,7 +7,7 @@ botCache.arguments.set("role", {
     const [id] = parameters;
     if (!id) return;
 
-    const guild = message.guild();
+    const guild = cache.guilds.get(message.guildID);
     if (!guild) return;
 
     const roleID = id.startsWith("<@&") ? id.substring(3, id.length - 1) : id;
