+++ src/commands/add.ts
@@ -1,6 +1,5 @@
 // This command is intentionally done in an un-optimized way. This command is only to show you how to await a users response.
-import { botCache } from "../../mod.ts";
-import { sendMessage } from "../../deps.ts";
+import { botCache, sendMessage } from "../../deps.ts";
 import { needMessage } from "../utils/collectors.ts";
 import { Embed } from "../utils/Embed.ts";
 import { sendEmbed } from "../utils/helpers.ts";
