+++ src/commands/help.ts
@@ -1,5 +1,4 @@
-import { botCache } from "../../mod.ts";
-import { sendMessage } from "../../deps.ts";
+import { botCache, sendMessage } from "../../deps.ts";
 import { translate } from "../utils/i18next.ts";
 import { Embed } from "../utils/Embed.ts";
