+++ src/commands/ping.ts
@@ -1,6 +1,5 @@
 // This file is meant to show how you can create multiple commands in the same file if you wish.
-import { botCache } from "../../mod.ts";
-import { cache, sendMessage } from "../../deps.ts";
+import { botCache, cache, sendMessage } from "../../deps.ts";
 
 botCache.commands.set(`ping`, {
   name: `ping`,
