+++ src/events/debug.ts
@@ -1,4 +1,4 @@
-import { botCache } from "../../mod.ts";
+import { botCache } from "../../deps.ts";
 
 botCache.eventHandlers.debug = function (data) {
   // console.warn(data);
