+++ src/inhibitors/nsfw.ts
@@ -1,5 +1,4 @@
-import { cache } from "../../deps.ts";
-import { botCache } from "../../mod.ts";
+import { botCache, cache } from "../../deps.ts";
 
 botCache.inhibitors.set("nsfw", async function (message, command, guild) {
   // If this command does not need nsfw the inhibitor returns false so the command can run
