+++ src/permissionLevels/admin.ts
@@ -1,6 +1,5 @@
-import { botCache } from "../../mod.ts";
+import { botCache, memberIDHasPermission } from "../../deps.ts";
 import { PermissionLevels } from "../types/commands.ts";
-import { memberIDHasPermission } from "../../deps.ts";
 
 // The member using the command must be an admin. (Required ADMIN server perm.)
 botCache.permissionLevels.set(
