+++ src/permissionLevels/moderator.ts
@@ -1,6 +1,5 @@
-import { botCache } from "../../mod.ts";
 import { PermissionLevels } from "../types/commands.ts";
-import { memberIDHasPermission } from "../../deps.ts";
+import { botCache, memberIDHasPermission } from "../../deps.ts";
 
 // The member using the command must be a moderator. (Usually has MANAGE_GUILD perm)
 botCache.permissionLevels.set(
