+++ src/tasks/collectors.ts
@@ -1,5 +1,5 @@
 // This task will help remove un-used collectors to help keep our cache optimized.
-import { botCache } from "../../mod.ts";
+import { botCache } from "../../deps.ts";
 import { Milliseconds } from "../utils/constants/time.ts";
 
 botCache.tasks.set(`collectors`, {
