+++ src/tasks/sweeper.ts
@@ -1,6 +1,5 @@
-import { botCache } from "../../mod.ts";
 import { Milliseconds } from "../utils/constants/time.ts";
-import { cache, cacheHandlers } from "../../deps.ts";
+import { botCache, cache, cacheHandlers } from "../../deps.ts";
 
 const MESSAGE_LIFETIME = Milliseconds.MINUTE * 10;
 const MEMBER_LIFETIME = Milliseconds.MINUTE * 30;
