+++ src/utils/collectors.ts
@@ -1,18 +1,13 @@
-import type {
-  Message,
-  MessageReactionUncachedPayload,
-  ReactionPayload,
-} from "../../deps.ts";
-import type {
+import {
   MessageCollectorOptions,
   ReactionCollectorOptions,
   CollectMessagesOptions,
   CollectReactionsOptions,
 } from "../types/collectors.ts";
-
-import { botID } from "../../deps.ts";
+import { botCache, botID, Message,
+  MessageReactionUncachedPayload,
+  ReactionPayload, } from "../../deps.ts";
 import { Milliseconds } from "./constants/time.ts";
-import { botCache } from "../../mod.ts";
 
 export async function needMessage(
   memberID: string,
